pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chat-sud"
version = "0.1.0"
description = "Privacy-first therapeutic support system for substance use recovery."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "numpy>=1.26.0",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
retrieval = [
  "faiss-cpu>=1.8.0",
  "rank-bm25>=0.2.2",
  "sentence-transformers>=3.0.0",
]
training = [
  "accelerate>=0.33.0",
  "bitsandbytes>=0.43.0",
  "datasets>=2.20.0",
  "peft>=0.12.0",
  "transformers>=4.44.0",
  "trl>=0.10.0",
]
pdf = [
  "pypdf>=4.3.0",
  "unstructured[pdf]>=0.15.0",
]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.3.0",
  "ruff>=0.6.0",
]

[project.scripts]
chat-sud-api = "chat_sud.api:run"
chat-sud-cli = "chat_sud.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["chat_sud"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

